Cosecant

Base.Math.csc — Function

csc(x)

Compute the cosecant of x, where x is in radians.

Base.Math.cscd — Function

cscd(x)

Compute the cosecant of x, where x is in degrees.

Base.Math.csch — Function

csch(x)

Compute the hyperbolic cosecant of x.

Base.Math.acsc — Function

acsc(x)

Compute the inverse cosecant of x, where the output is in radians.

Base.Math.acscd — Function

acscd(x)

Compute the inverse cosecant of x, where the output is in degrees. If x is a matrix, x needs to be a square matrix.

Julia 1.7

Matrix arguments require Julia 1.7 or later.

Base.Math.acsch — Function

acsch(x)

Compute the inverse hyperbolic cosecant of x.