Secant

Base.Math.sec — Function

sec(x)

Compute the secant of x, where x is in radians.

Base.Math.secd — Function

secd(x)

Compute the secant of x, where x is in degrees.

Base.Math.sech — Function

sech(x)

Compute the hyperbolic secant of x.

Base.Math.asec — Function

asec(x)

Compute the inverse secant of x, where the output is in radians.

Base.Math.asecd — Function

asecd(x)

Compute the inverse secant of x, where the output is in degrees. If x is a matrix, x needs to be a square matrix.

Julia 1.7

Matrix arguments require Julia 1.7 or later.

Base.Math.asech — Function

asech(x)

Compute the inverse hyperbolic secant of x.