acsc

Compute the principal value of the arc cosecant of x, return in radians.

Base.Math.acsc — Function

acsc(x)

Compute the inverse cosecant of x, where the output is in radians.

acsc(A::AbstractMatrix)

Compute the inverse matrix cosecant of A.

Methods

julia> methods(acsc, (Any,), [Base, Base.Math, Base.MathConstants, Base.MPFR])# 1 method for generic function "acsc" from Base.Math:
 [1] acsc(y::Number)
     @ special/trig.jl:1157

Examples

Real Numbers

julia> acsc(Inf)
0.0

julia> acsc(-Inf)
-0.0

julia> acsc(1.0) / pi
0.5

julia> acsc(0)
ERROR: DomainError with Inf:
asin(x) is not defined for |x| > 1.
Stacktrace:
[...]

Complex

julia> acsc(Inf+0im)
0.0 - 0.0im

Tips

See Also

Extended Inputs

Matrix

With Array like input:

julia> methods(acsc, (Any,), [LinearAlgebra])# 3 methods for generic function "acsc" from Base.Math:
 [1] acsc(J::UniformScaling)
     @ LinearAlgebra /opt/hostedtoolcache/julia/1.11.5/x64/share/julia/stdlib/v1.11/LinearAlgebra/src/uniformscaling.jl:173
 [2] acsc(D::Diagonal)
     @ LinearAlgebra /opt/hostedtoolcache/julia/1.11.5/x64/share/julia/stdlib/v1.11/LinearAlgebra/src/diagonal.jl:802
 [3] acsc(A::AbstractMatrix{T}) where T
     @ LinearAlgebra /opt/hostedtoolcache/julia/1.11.5/x64/share/julia/stdlib/v1.11/LinearAlgebra/src/dense.jl:1365

Tech Notes

Implemented in terms of asin: acsc(x) = inv(asin(x))

Version History

Introduced in Julia v1.0 (2018)

External Links

🔗DLMF: 4.23.7

🔗Inverse Cosecant - MathWorld

🔗Inverse trigonometric functions - Wikipedia