Cosine

Base.cos — Function

cos(x)

Compute cosine of x, where x is in radians.

See also cosd, cospi, sincos, cis.

Base.Math.cosd — Function

cosd(x)

Compute cosine of x, where x is in degrees. If x is a matrix, x needs to be a square matrix.

Julia 1.7

Matrix arguments require Julia 1.7 or later.

Base.Math.cospi — Function

cospi(x)

Compute $\cos(\pi x)$ more accurately than cos(pi*x), especially for large x.

Base.cosh — Function

cosh(x)

Compute hyperbolic cosine of x.

Base.acos — Function

acos(x)

Compute the inverse cosine of x, where the output is in radians

Base.Math.acosd — Function

acosd(x)

Compute the inverse cosine of x, where the output is in degrees. If x is a matrix, x needs to be a square matrix.

Julia 1.7

Matrix arguments require Julia 1.7 or later.

Base.acosh — Function

acosh(x)

Compute the inverse hyperbolic cosine of x.

Base.Math.cosc — Function

cosc(x)

Compute $\cos(\pi x) / x - \sin(\pi x) / (\pi x^2)$ if $x \neq 0$, and $0$ if $x = 0$. This is the derivative of sinc(x).