asecd

Compute the principal value of the arc secant of x, return in degrees.

Base.Math.asecd — Function

asecd(x)

Compute the inverse secant of x, where the output is in degrees. If x is a matrix, x needs to be a square matrix.

Julia 1.7

Matrix arguments require Julia 1.7 or later.

Methods

julia> methods(asecd, (Any,), [Base, Base.Math, Base.MathConstants, Base.MPFR])# 1 method for generic function "asecd" from Base.Math:
 [1] asecd(x)
     @ special/trig.jl:1287

Examples

Real Numbers

julia> asecd(1.0)
0.0

julia> asecd(sqrt(2))
45.00000000000001

julia> asecd(2.0)
60.00000000000001

julia> asecd(Inf)
90.0

julia> asecd(-1.0)
180.0

Complex

julia> asecd(1+0im)
0.0 + 0.0im

Tips

See Also

Extended Inputs

Tech Notes

Implemented in terms of asec: asecd(x) = rad2deg(asec(x))

Version History

Introduced in Julia v1.0 (2018)

External Links