Cotangent

Base.Math.cot — Function

cot(x)

Compute the cotangent of x, where x is in radians.

Base.Math.cotd — Function

cotd(x)

Compute the cotangent of x, where x is in degrees.

Base.Math.coth — Function

coth(x)

Compute the hyperbolic cotangent of x.

Base.Math.acot — Function

acot(x)

Compute the inverse cotangent of x, where the output is in radians.

Base.Math.acotd — Function

acotd(x)

Compute the inverse cotangent of x, where the output is in degrees. If x is a matrix, x needs to be a square matrix.

Julia 1.7

Matrix arguments require Julia 1.7 or later.

Base.Math.acoth — Function

acoth(x)

Compute the inverse hyperbolic cotangent of x.